feat: moved zustand stores to seperate folder
Moved zustand stores to seperate folder, also added new dashboard store
which keeps the state of the dashboard
