+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/components/DetailedOrgView.tsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from "react";
 import { Typography, Box } from "@mui/material";
 
 import { PersonalContact } from "../models/mainContentTypes";
-import { useAppStore } from "../../../hooks/Appstore";
 import OfficialContactFieldTable from "./contacts/NotificationContactTable";
 import ERRolesTable from "./ERRolesTable";
 import { RoleDetails } from "./RoleDetails";
@@ -11,6 +10,7 @@ import { officialContactsBoxStyle } from "../styles/DetailedOrgView.styles";
 import ContactsTable from "./contacts/ContactsTable";
 import { useOrgDetails } from "../../../hooks/hooks";
 import { SelectedOrg } from "../../../models/models";
+import { useAppStore } from "../../../stores/Appstore";
 
 interface DetailedOrgViewProps {
   selectedOrg: SelectedOrg;
