+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/components/organizations/OrganizationList.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { Alert, Typography, Skeleton } from "@mui/material";
 import { OrganizationCard } from "./OrganizationCard";
 import { useOrgSearch } from "../../../../hooks/hooks";
-import { useAppStore } from "../../../../hooks/Appstore";
 import { ErrorAlert } from "../ErrorAlert";
 import { Organization, SelectedOrg } from "../../../../models/models";
+import { useAppStore } from "../../../../stores/Appstore";
 
 interface OrganizationListProps {
   setSelectedOrg: (SelectedOrg: SelectedOrg) => void;
@@ -45,7 +45,10 @@ export const OrganizationList: React.FC<OrganizationListProps> = ({
 
   // Default case: render organizations
   return (
-    <div className={`org-list`} style={{ overflowY: "auto", maxHeight: "80vh" }}>
+    <div
+      className={`org-list`}
+      style={{ overflowY: "auto", maxHeight: "80vh" }}
+    >
       {organizations
         .filter((org) => {
           // filter out subunits if parent is already included
