+++ frontend/altinn-support-dashboard.client/src/pages/NewOrganizationPage.tsx
@@ -1,5 +1,5 @@
 import OrganizationCreationComponent from "../components/OrganizationCreation/OrganizationCreationComponent";
-import { useAppStore } from "../hooks/Appstore";
+import { useAppStore } from "../stores/Appstore";
 
 const NewOrganizationPage: React.FC = () => {
   const environment = useAppStore.getState().environment;
