+++ frontend/altinn-support-dashboard.client/src/stores/Appstate.ts
