+++ frontend/altinn-support-dashboard.client/src/stores/Appstore.ts
