+++ frontend/altinn-support-dashboard.client/src/stores/Dashboardstate.ts
@@ -1,0 +1,8 @@
+import { SelectedOrg } from "../models/models";
+
+export interface DashboardState {
+  query: string;
+  selectedOrg: SelectedOrg | null;
+  setQuery: (q: string) => void;
+  setSelectedOrg: (org: SelectedOrg | null) => void;
+}
